finetunning/simpleimdb/requirements.txt
python-dotenv
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langcodes==3.4.0
langsmith==0.1.55
language_data==1.2.0
libclang==18.1.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.25.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
posthog==3.5.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
tf-keras

